probleme export OK
