+++ includes/minishell.h
@@ -125,6 +125,7 @@ int					execute_builtin(t_minishell_context *minishell_context,
 						t_command *command, bool is_in_pipeline);
 int					env(t_command *current_command);
 int					export(t_command *command);
+void				display_sorted_exportables_variables(char **env);
 int					pwd(t_command *command);
 
 // EXIT STATUS
